Made implicit typecasts explicit to eliminate compiler warnings

+++ srtp/CryptoContext.cpp
@@ -398,7 +398,7 @@ bool CryptoContext::checkReplay(uint16_t newSeq)
         }
 
         delta = -delta;
-        int idx = delta / 64;
+        int idx = (int)(delta / 64);
         uint64_t bit = 1UL << (delta % 64);
         if ((replay_window[idx] & bit) == bit) {
             return false;  /* Packet already received ! */
@@ -447,7 +447,7 @@ void CryptoContext::update(uint16_t newSeq)
     }
     else {
         delta = -delta;
-        int idx = delta / 64;
+        int idx = (int)(delta / 64);
         uint64_t bit = 1UL << (delta % 64);
         replay_window[idx] |= bit;
     }
