+++ zrtp/zrtpB64Encode.c
@@ -46,7 +46,7 @@ int base64_encode_block(const uint8_t *plaintext_in, int length_in, char* code_o
             {
                 state_in->result = result;
                 state_in->step = step_A;
-                return codechar - code_out;
+                return (int)(codechar - code_out);
             }
             fragment = *plainchar++;
             result = (fragment & 0x0fc) >> 2;
@@ -57,7 +57,7 @@ int base64_encode_block(const uint8_t *plaintext_in, int length_in, char* code_o
             {
                 state_in->result = result;
                 state_in->step = step_B;
-                return codechar - code_out;
+                return (int)(codechar - code_out);
             }
             fragment = *plainchar++;
             result |= (fragment & 0x0f0) >> 4;
@@ -68,7 +68,7 @@ int base64_encode_block(const uint8_t *plaintext_in, int length_in, char* code_o
             {
                 state_in->result = result;
                 state_in->step = step_C;
-                return codechar - code_out;
+                return (int)(codechar - code_out);
             }
             fragment = *plainchar++;
             result |= (fragment & 0x0c0) >> 6;
@@ -111,6 +111,6 @@ int base64_encode_blockend(char* code_out, base64_encodestate* state_in)
     if (state_in->lineLength > 0)
         *codechar++ = '\n';
 
-    return codechar - code_out;
+    return (int)(codechar - code_out);
 }
